updated version

+++ pr1/app1/models.py
@@ -6,8 +6,9 @@ class Movie(models.Model):
     movieName = models.TextField()
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
-    rating = models.FloatField(default=0)
+    rating = models.DecimalField(default=0, max_digits=3, decimal_places=1)
     owner = models.ForeignKey(User, on_delete=models.CASCADE)
+
 
     def __str__(self):
         return self.movieName
